Collect DataFrame reduction tess
